feat: verify Gazebo Fortress → ROS 2 → mock GPS noise pipeline in Docker
End-to-end test confirmed all Layers 1 and 2b of the Simulink bridge:
- Layer 1: /model/iris_uav/odometry at 49.8 Hz, IMU topic present
- Layer 2b: /fix at 4.995 Hz with realistic GPS noise (lat/lon confirmed)
- Layer 3: accuracy_verify.py --use-fix wiring confirmed

Key fixes found during testing:
- ros-humble-ros-gz-bridge links against ignition-transport11 (Fortress),
  not gz-transport13 (Harmonic); switched Dockerfile to ignition-fortress
- Added IGN_IP=127.0.0.1 to force unicast (Docker blocks multicast)
- Start bridge before gz sim to avoid missed discovery announcements
- Bridge message types: ignition.msgs.* not gz.msgs.*
- mock_fix_publisher position_covariance: explicit float() cast required
- Added robotx_docker_test.sdf: self-contained Fortress world (no ardupilot_gazebo)
- Added coast_waves stub model for offline/Docker use without VRX
- docker/diag.sh: diagnostic helper for transport debugging
- Updated docs/simulink_bridge/README.md with live test results

+++ simulation/mock_fix_publisher.py
@@ -134,7 +134,7 @@ class MockFixPublisher(Node):
         msg.altitude  = 0.0   # AGL altitude not needed for 2D buoy projection
         # Diagonal covariance: sigma^2 in lat, lon, alt (metres^2)
         cov = sigma_m ** 2
-        msg.position_covariance = [cov, 0, 0,  0, cov, 0,  0, 0, cov * 4]
+        msg.position_covariance = [float(cov), 0.0, 0.0,  0.0, float(cov), 0.0,  0.0, 0.0, float(cov * 4)]
         msg.position_covariance_type = NavSatFix.COVARIANCE_TYPE_DIAGONAL_KNOWN
         self._pub.publish(msg)
